+++ src/lib/config/core.js
@@ -4,10 +4,11 @@ import possibleErrorsRules from './rules/possible-errors';
 import strictModeRules from './rules/strict-mode';
 import stylisticIssuesRules from './rules/stylistic-issues';
 import variablesRules from './rules/variables';
+import xoRules from './rules/xo';
 
 export default {
     plugins: [
-        'objects'
+        'xo'
     ],
     ecmaFeatures: {
         ecmaVersion: 5
@@ -17,6 +18,7 @@ export default {
         bestPracticesRules,
         strictModeRules,
         variablesRules,
-        stylisticIssuesRules
+        stylisticIssuesRules,
+        xoRules
     )
 };